(1248) Split up userAccessModel
